refine channelActivity, showAttachment from MessageListView
