+++ sample/src/main/java/io/getstream/chat/example/ChannelActivity.java
@@ -12,26 +12,18 @@ import androidx.lifecycle.ViewModelProviders;
 import com.getstream.sdk.chat.StreamChat;
 import com.getstream.sdk.chat.model.Attachment;
 import com.getstream.sdk.chat.model.Channel;
-import com.getstream.sdk.chat.model.ModelType;
 import com.getstream.sdk.chat.rest.Message;
 import com.getstream.sdk.chat.rest.core.Client;
 import com.getstream.sdk.chat.utils.Constant;
 import com.getstream.sdk.chat.utils.PermissionChecker;
-import com.getstream.sdk.chat.utils.frescoimageviewer.ImageViewer;
 import com.getstream.sdk.chat.view.Dialog.ReactionDialog;
 import com.getstream.sdk.chat.view.MessageInputView;
 import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.Dialog.MoreActionDialog;
-import com.getstream.sdk.chat.view.activity.AttachmentActivity;
-import com.getstream.sdk.chat.view.activity.AttachmentDocumentActivity;
-import com.getstream.sdk.chat.view.activity.AttachmentMediaActivity;
 import com.getstream.sdk.chat.viewmodel.ChannelViewModel;
 import com.getstream.sdk.chat.viewmodel.ChannelViewModelFactory;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import io.getstream.chat.example.databinding.ActivityChannelBinding;
 
@@ -70,13 +62,23 @@ public class ChannelActivity extends AppCompatActivity
         viewModel = ViewModelProviders.of(this,
                 new ChannelViewModelFactory(this.getApplication(), channel)
         ).get(ChannelViewModel.class);
-        viewModel.getChannel().setReactionTypes(reactionTypes);
+        // set custom Reaction Emojis
+        viewModel.getChannel().setReactionTypes(new HashMap<String, String>() {
+            {
+                put("like", "\uD83D\uDC4D");
+                put("love", "\u2764\uFE0F");
+                put("haha", "\uD83D\uDE02");
+                put("wow", "\uD83D\uDE32");
+                put("sad", " \uD83D\uDE41");
+                put("angry", "\uD83D\uDE21");
+                put("cheeky", "\uD83D\uDE1B");
+            }
+        });
         // connect the view model
         binding.channelHeader.setViewModel(viewModel, this);
         binding.channelHeader.setOnBackClickListener(v -> finish());
 
-        MyMessageViewHolderFactory factory = new MyMessageViewHolderFactory();
-        binding.messageList.setViewHolderFactory(factory);
+        binding.messageList.setViewHolderFactory(new MyMessageViewHolderFactory());
         binding.messageList.setMessageClickListener(this);
         binding.messageList.setMessageLongClickListener(this);
         binding.messageList.setAttachmentClickListener(this);
@@ -86,6 +88,13 @@ public class ChannelActivity extends AppCompatActivity
         binding.messageInput.setOpenCameraViewListener(this);
         // set the viewModel data for the activity_channel.xml layout
         binding.setViewModel(viewModel);
+    }
+
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        binding.messageInput.progressCapturedMedia(requestCode, resultCode, data);
     }
 
     @Override
@@ -108,98 +117,11 @@ public class ChannelActivity extends AppCompatActivity
     }
 
     @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        binding.messageInput.progressCapturedMedia(requestCode, resultCode, data);
-    }
-
-    @Override
     public void onMessageClick(Message message, int position) {
         ReactionDialog reactionDialog = new ReactionDialog(this,
                 viewModel.getChannel(), message, position, binding.messageList, binding.messageList.getStyle());
         reactionDialog.show();
     }
-
-    @Override
-    public void onAttachmentClick(Message message, Attachment attachment) {
-        String url = null;
-        String type = null;
-        switch (attachment.getType()) {
-            case ModelType.attach_file:
-                loadFile(attachment);
-                return;
-            case ModelType.attach_image:
-                if (attachment.getOgURL() != null) {
-                    url = attachment.getOgURL();
-                    type = ModelType.attach_link;
-                } else {
-                    List<String> imageUrls = new ArrayList<>();
-                    for (Attachment a : message.getAttachments()) {
-                        imageUrls.add(a.getImageURL());
-                    }
-                    int position = message.getAttachments().indexOf(attachment);
-
-                    new ImageViewer.Builder<>(this, imageUrls)
-                            .setStartPosition(position)
-                            .show();
-                    return;
-                }
-                break;
-            case ModelType.attach_video:
-                url = attachment.getTitleLink();
-                break;
-            case ModelType.attach_giphy:
-                url = attachment.getThumbURL();
-                break;
-            case ModelType.attach_product:
-                url = attachment.getUrl();
-                break;
-        }
-        if (type == null) type = attachment.getType();
-        Intent intent = new Intent(this, AttachmentActivity.class);
-        intent.putExtra("type", type);
-        intent.putExtra("url", url);
-        startActivity(intent);
-    }
-
-    // region Load File
-    private void loadFile(Attachment attachment) {
-        // Media
-        if (attachment.getMime_type().contains("audio") ||
-                attachment.getMime_type().contains("video")) {
-            Intent intent = new Intent(this, AttachmentMediaActivity.class);
-            intent.putExtra("type", attachment.getMime_type());
-            intent.putExtra("url", attachment.getAssetURL());
-            startActivity(intent);
-            return;
-        }
-
-        // Office
-        if (attachment.getMime_type().equals("application/msword") ||
-                attachment.getMime_type().equals(ModelType.attach_mime_txt) ||
-                attachment.getMime_type().equals(ModelType.attach_mime_pdf) ||
-                attachment.getMime_type().contains("application/vnd")) {
-
-            Intent intent = new Intent(this, AttachmentDocumentActivity.class);
-            intent.putExtra("url", attachment.getAssetURL());
-            startActivity(intent);
-        }
-    }
-    // endregion
-
-
-    Map<String, String> reactionTypes = new HashMap<String, String>() {
-        {
-            put("like", "\uD83D\uDC4D");
-            put("love", "\u2764\uFE0F");
-            put("haha", "\uD83D\uDE02");
-            put("wow", "\uD83D\uDE32");
-            put("sad", " \uD83D\uDE41");
-            put("angry", "\uD83D\uDE21");
-            put("cheeky", "\uD83D\uDE1B");
-        }
-    };
-
 
     @Override
     public void onMessageLongClick(Message message) {
@@ -209,4 +131,10 @@ public class ChannelActivity extends AppCompatActivity
                 binding.messageList.getStyle());
         moreActionDialog.show();
     }
+
+    @Override
+    public void onAttachmentClick(Message message, Attachment attachment) {
+        binding.messageList.showAttachment(message, attachment);
+    }
+
 }
